+++ src/languageservice/yamlLanguageService.ts
@@ -222,7 +222,7 @@ export function getLanguageService(params: {
       }
       yamlValidation.configure(settings);
       hover.configure(settings);
-      completer.configure(settings);
+      completer.configure(settings, params.yamlSettings);
       formatter.configure(settings);
       yamlCodeActions.configure(settings);
     },